Enhancement #32: Resource fetching done
