+++ Project Code/ARSchoolBook/.idea/misc.xml
@@ -4,7 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Study Materials/11. Spring 2021/CSE 499A/Project/Github Repo/Group-01/Project Code/ARSchoolBook/app/src/main/res/layout/activity_class_selection.xml" value="0.24895833333333334" />
-        <entry key="..\:/Study Materials/11. Spring 2021/CSE 499A/Project/Github Repo/Group-01/Project Code/ARSchoolBook/app/src/main/res/layout/activity_ge_home.xml" value="0.24895833333333334" />
+        <entry key="..\:/Study Materials/11. Spring 2021/CSE 499A/Project/Github Repo/Group-01/Project Code/ARSchoolBook/app/src/main/res/layout/activity_ge_home.xml" value="0.264" />
         <entry key="..\:/Study Materials/11. Spring 2021/CSE 499A/Project/Github Repo/Group-01/Project Code/ARSchoolBook/app/src/main/res/layout/activity_home.xml" value="0.24895833333333334" />
         <entry key="..\:/Study Materials/11. Spring 2021/CSE 499A/Project/Github Repo/Group-01/Project Code/ARSchoolBook/app/src/main/res/layout/chapter_item_view.xml" value="0.24895833333333334" />
         <entry key="..\:/Study Materials/11. Spring 2021/CSE 499A/Project/Github Repo/Group-01/Project Code/ARSchoolBook/app/src/main/res/layout/class_item_view.xml" value="0.24895833333333334" />
